minor refactoring of ingredient-functions
